move cert authority settings under topology, add tests and add keysize to node
