+++ tests/01-smoke/09-external-ca.clab.yml
@@ -1,0 +1,18 @@
+# Copyright 2020 Nokia
+# Licensed under the BSD 3-Clause License.
+# SPDX-License-Identifier: BSD-3-Clause
+
+name: external-ca
+  
+topology:
+  settings:
+    certificate-authority:
+      cert: rootCACert.pem
+      key: rootCAKey.pem
+  nodes:
+    l1:
+      kind: linux
+      image: alpine:3
+      cmd: ash -c "sleep 9999"
+      certificate:
+        issue: true
