+++ types/topology.go
@@ -12,6 +12,7 @@ type Topology struct {
 	Kinds    map[string]*NodeDefinition `yaml:"kinds,omitempty"`
 	Nodes    map[string]*NodeDefinition `yaml:"nodes,omitempty"`
 	Links    []*links.LinkDefinition    `yaml:"links,omitempty"`
+	Settings *Settings                  `yaml:"settings"`
 }
 
 func NewTopology() *Topology {
